+++ lib/src/button.dart
@@ -1,22 +1,23 @@
 part of '../tab_navigation_bar.dart';
 
 class _TabNavigationButton extends StatelessWidget {
-  final VoidCallback onPressed;
+  final VoidCallback? onPressed;
   final AnimationController controller;
-  final IconData selectedIcon;
-  final IconData unselectedIcon;
+  final IconData? selectedIcon;
+  final IconData? unselectedIcon;
   final int selectedIndex;
   final int index;
-  final Color color;
+  final Color? color;
   final double iconSize;
-  final Color inactiveColor;
-  final Color barColor;
+  final Color? inactiveColor;
+  final Color? barColor;
   final double bottomPadding;
-  final double barHeight;
+  final double height;
+  final int itemCount;
 
   const _TabNavigationButton({
-    super.key,
-    required this.onPressed,
+    required this.itemCount,
+    this.onPressed,
     required this.controller,
     required this.selectedIcon,
     required this.unselectedIcon,
@@ -27,20 +28,20 @@ class _TabNavigationButton extends StatelessWidget {
     required this.inactiveColor,
     required this.barColor,
     required this.bottomPadding,
-    required this.barHeight,
+    required this.height,
   });
 
   @override
   Widget build(BuildContext context) {
     final MediaQueryData mediaQueryData = MediaQuery.of(context);
     final double deviceWidth = mediaQueryData.size.width;
-    final double maxElementWidth = deviceWidth / 4;
+    final double maxElementWidth = deviceWidth / itemCount;
     final isSelected = selectedIndex == index;
     return GestureDetector(
       onTap: onPressed,
       child: Container(
         width: maxElementWidth,
-        height: barHeight,
+        height: height,
         color: barColor,
         child: Icon(
           isSelected ? selectedIcon : unselectedIcon,