+++ lib/src/item.dart
@@ -1,12 +1,14 @@
 part of '../tab_navigation_bar.dart';
 
 class TabNavigationItem {
-  final IconData icon;
+  final IconData? icon;
 
   final IconData? activeIcon;
 
   const TabNavigationItem({
+    this.icon,
     this.activeIcon,
-    required this.icon,
   });
+
+  bool get _isValidItem => icon != null || activeIcon != null;
 }