+++ CHANGELOG.md
@@ -1,1 +1,7 @@
+## 1.0.2
+* Indicator padding bug fixed
+
+## 1.0.1
+* Change items limitation 4 to 5
+
 ## 1.0.0